requirements.txt
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
langchain-ollama>=0.1.0
langgraph>=0.1.0
faiss-cpu>=1.8.0
tiktoken>=0.7.0
python-dotenv>=1.0.0